Migrate changes to pokemon components.
