+++ src/styles/baseline.scss
@@ -35,6 +35,10 @@ h2 {
   padding-bottom: 15px;
 }
 
+h4, h5 {
+  text-transform: capitalize;
+}
+
 .datetime {
   color: #fff;
   background: #6d7c82;
@@ -93,5 +97,5 @@ span.api-source-graphql {
 }
 
 span.api-source-sales {
-  background: rgb(212, 96, 0);  
+  background: rgb(212, 96, 0);
 }
