+++ src/styles/pokemon.scss
@@ -159,7 +159,7 @@ div.docs-refs {
   >.inside {
     display: block;
     width: 100%;
-    max-width: 800px;
+    max-width: 1000px;
     float: left;
   }
   .option-rows {
